requirements.txt
Django==1.10
django-bower==5.2.0
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.7
psycopg2==2.6.2
six==1.10.0
